Add in limited support for Internal Areas
The code as is does not support printing or defining internal areas.
This patch adds basic support for doing this.

The contents of the internal area are defined by a list of hex numbers as
shown below:

InternalUse:
  data: 01 02 03 04 ab cd ef

The size of the internal area is set to the recommended 72 by default.
It can be changed by using the --internal-area-size command line arg.
The data given  will be padded with 0xff or truncated to fit this
length.  Note that byte 0 of the internal area is mandated to be a '1'
(see section 9 of the FRU spec), and that is included in the size.

When dumping, since there is no length field in the internal area, it
just takes the start of the next section (if present) as the size to
dump.