Add support for PowerSupplyInformation record in MultiRecord area
There was no nice way to represent 12-bit fields as FixedField-s because
the bits appear to be disjoint with the bit and byte orders that are
used: MSB first for bits, LSB first for bytes.
Here is an illustration of two fields, 4-bit field A and 12-bit field B,
where field A is located in the most significant bits of a little-endian
2-byte memory area:
    B7 B6 B5 B4 B3 B2 B1 B0  A3 A2 A1 A0 B11 B10 B9 B8
I could not find any way to unpack A and B using bitstruct.

So, I had to resort to using _mergeBitfield=True.
That flag reverses the order of all bytes in a record, thus, converting
any multi-byte little-endian fields to big-endian.  That way, the bytes
and bits have the same ordering and the sub-fields become contiguous:
    A3 A2 A1 A0 B11 B10 B9 B8  B7 B6 B5 B4 B3 B2 B1 B0
bitstruct can handle this bit order naturally.

As a result of using _mergeBitfield,  bytes are defined in the opposite
order comparing to Table 18-1.